+++ app/coins/btcHolidays.js
@@ -4,7 +4,7 @@ module.exports = {
 		{
 			name: "Bitcoin Pizza Day",
 			date: "2010-05-22",
-			desc: "On this day in 2010, Lazlo Hanyecz, paid Jeremy Sturdivant 10,000 BTC for two pizzas costing ~41 USD. The famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a> was one of the earliest real-world BTC transactions and is celebrated as helping start the Bitcoin economy."
+			desc: "On this day in 2010, Laszlo Hanyecz, paid Jeremy Sturdivant 10,000 BTC for two pizzas costing ~41 USD. The famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a> was one of the earliest real-world BTC transactions and is celebrated as helping start the Bitcoin economy."
 		},
 		{
 			name: "SegWit Lock-In Day",
@@ -59,7 +59,7 @@ module.exports = {
 		{
 			name: "Proof of Keys Day",
 			date: "2019-01-03",
-			desc: "On this day in 2019, to celebrate Bitcoin's 10th Birthday, Trace Mayer started an annual Proof-of-Keys Celebration. Participats declare monetary sovereignty by withdrawing their BTC into their own custory. Not your keys, not your bitcoin! You can read Trace Mayer's original tweet <a href='https://twitter.com/TraceMayer/status/1071870548421066753'>here</a>."
+			desc: "On this day in 2019, to celebrate Bitcoin's 10th Birthday, Trace Mayer started an annual Proof-of-Keys Celebration. Participats declare monetary sovereignty by withdrawing their BTC into their own custody. Not your keys, not your bitcoin! You can read Trace Mayer's original tweet <a href='https://twitter.com/TraceMayer/status/1071870548421066753'>here</a>."
 		},
 		{
 			name: "Genesis Transaction Day",
@@ -69,7 +69,7 @@ module.exports = {
 		{
 			name: "Lightning Pizza Day",
 			date: "2018-02-25",
-			desc: "On this day in 2018, eight years after buying pizza with the famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a>, Lazlo Hanyecz returned to buy pizza over the Lightning Network."
+			desc: "On this day in 2018, eight years after buying pizza with the famous <a href='./tx/a1075db55d416d3ca199f55b6084e2115b9345e16c5cf302fc80e9d5fbf5d48d@57043'>\"Bitcoin Pizza\" transaction</a>, Laszlo Hanyecz returned to buy pizza over the Lightning Network."
 		},
 		{
 			name: "Mt. Gox Death Day",
@@ -130,6 +130,11 @@ module.exports = {
 			name: "Goodbye Satoshi Day",
 			date: "2010-12-12",
 			desc: "On this day in 2010, Satoshi sent his final message on the BitcoinTalk forum. Then, he logged off, never to return."
+		},
+		{
+			name: "Satoshi's Birthday",
+			date: "1975-04-05",
+			desc: "Satoshi Nakamoto listed their birthday as April 5, 1975. Many have speculated about symbolic significance of this date, with some identifying April 5 as the date that FDR signed <a href='https://en.wikipedia.org/wiki/Executive_Order_6102'>Executive Order 6102</a>, and 1975 as the year that Americans were legally allowed to old gold again, after Nixon's 1971 ban."
 		}
 	]
 };